rebrand executable, add manga stuff, etc
