+++ build.gradle
@@ -18,7 +18,7 @@ application {
     mainClass.set("com.hrothwell.anime.Main")
 }
 
-run{
+run {
     standardInput = System.in
 }
