--- native-image-items/build-results/anime.build_artifacts.txt
@@ -1,6 +1,0 @@
-[EXECUTABLE]
-anime.exe
-
-[JDK_LIB]
-sunmscapi.dll
-
