+++ native-image-items/build-results/mal.build_artifacts.txt
@@ -1,0 +1,6 @@
+[EXECUTABLE]
+mal.exe
+
+[JDK_LIB]
+sunmscapi.dll
+
