+++ src/main/kotlin/com/hrothwell/anime/commands/Suggest.kt
@@ -3,11 +3,11 @@ package com.hrothwell.anime.commands
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.parameters.options.default
 import com.github.ajalt.clikt.parameters.options.option
-import com.hrothwell.anime.client.MalAnimeClient
+import com.hrothwell.anime.client.MalClient
 import com.hrothwell.anime.util.AnimeUtil
 
 class Suggest : CliktCommand(
-  help = "Get anime suggestions!"
+  help = "Get anime suggestions! Manga suggestions not yet supported by MAL"
 ) {
 
   val limit by option(
@@ -19,7 +19,7 @@ class Suggest : CliktCommand(
 
   override fun run() {
     try {
-      val animeList = MalAnimeClient.getSuggestedAnime(limit.toInt())
+      val animeList = MalClient.getSuggestedAnime(limit.toInt())
       echo("Try: ${animeList.map { it.title }}")
       if (animeList.size == 1) {
         AnimeUtil.openAnime(animeList.first())
